+++ src/server/node_types.rs
@@ -113,6 +113,15 @@ impl<T: Attributes> fmt::Debug for Node<T> {
 }
 
 #[derive(Debug, Clone)]
+pub struct DataTypeNode {
+    pub requested_new_node_id: Option<ua::NodeId>,
+    pub parent_node_id: ua::NodeId,
+    pub reference_type_id: ua::NodeId,
+    pub browse_name: ua::QualifiedName,
+    pub attributes: ua::DataTypeAttributes,
+}
+
+#[derive(Debug, Clone)]
 pub struct ObjectNode {
     pub requested_new_node_id: Option<ua::NodeId>,
     pub parent_node_id: ua::NodeId,